Use a templated base class to consolidate TNC tests.
